lite_LLM/pyproject.toml
[project]
name = "project-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.72.7",
    "python-dotenv>=1.1.0",
]
